@@ -1,5 +1,5 @@
 import { Component, Inject, OnInit } from '@angular/core';
-import { FormBuilder, FormGroup, Validators } from '@angular/forms';
+import { UntypedFormBuilder, UntypedFormGroup, Validators } from '@angular/forms';
 import { MatDialog, MAT_DIALOG_DATA } from '@angular/material/dialog';
 import { ToastrService } from 'ngx-toastr';
 import { Brand } from '../../../models/brand';
@@ -11,9 +11,9 @@ import { BrandService } from '../../../services/brand.service';
   styleUrls: ['./brand-form.component.scss']
 })
 export class BrandFormComponent implements OnInit {
-  brandForm: FormGroup;
+  brandForm: UntypedFormGroup;
   formTitle: string;
-  constructor(@Inject(MAT_DIALOG_DATA) public data: Brand, private dialogRef: MatDialog, private brandService: BrandService, private toastr: ToastrService, private fb: FormBuilder) { }
+  constructor(@Inject(MAT_DIALOG_DATA) public data: Brand, private dialogRef: MatDialog, private brandService: BrandService, private toastr: ToastrService, private fb: UntypedFormBuilder) { }
 
   ngOnInit(): void {
     this.initializeForm();
