@@ -1,5 +1,5 @@
 import { Component, Inject, OnInit } from "@angular/core";
-import { FormBuilder, FormControl, FormGroup, Validators } from "@angular/forms";
+import { UntypedFormBuilder, FormControl, UntypedFormGroup, Validators } from "@angular/forms";
 import { MAT_DIALOG_DATA } from "@angular/material/dialog";
 import { ToastrService } from "ngx-toastr";
 import { Upload } from "src/app/core/models/uploads/upload";
@@ -21,7 +21,7 @@ import { ProductService } from "../../../services/product.service";
     styleUrls: ["./product-form.component.scss"]
 })
 export class ProductFormComponent implements OnInit {
-    productForm: FormGroup;
+    productForm: UntypedFormGroup;
     formTitle: string;
     brands: PaginatedResult<Brand>;
     brandParams = new BrandParams();
@@ -39,7 +39,7 @@ export class ProductFormComponent implements OnInit {
         private categoryService: CategoryService,
         private toastr: ToastrService,
         private warehouseService: WarehouseService,
-        private fb: FormBuilder
+        private fb: UntypedFormBuilder
     ) {}
 
     ngOnInit(): void {
