@@ -1,5 +1,5 @@
 import { Component, Inject, OnInit } from "@angular/core";
-import { FormGroup, FormBuilder, Validators } from "@angular/forms";
+import { UntypedFormGroup, UntypedFormBuilder, Validators } from "@angular/forms";
 import { MAT_DIALOG_DATA, MatDialog } from "@angular/material/dialog";
 import { ToastrService } from "ngx-toastr";
 import { RequestStatus } from "src/app/core/enums/RequestStatus";
@@ -20,7 +20,7 @@ import { EmployeeService } from "../../../services/employee.service";
   styleUrls: ['./attendance-log-form.component.scss']
 })
 export class AttendanceLogFormComponent implements OnInit {
-  formGroup: FormGroup;
+  formGroup: UntypedFormGroup;
   formTitle: string;
   employeesLookup: Employee[];
   constructor(
@@ -29,7 +29,7 @@ export class AttendanceLogFormComponent implements OnInit {
       private attendanceService: AttendanceRequestService,
       private employeeService: EmployeeService,
       private toastr: ToastrService,
-      private fb: FormBuilder,
+      private fb: UntypedFormBuilder,
       private authService: AuthService
   ) {}
 
