@@ -1,5 +1,5 @@
 import { Component, OnInit } from "@angular/core";
-import { FormGroup, FormBuilder, Validators } from "@angular/forms";
+import { UntypedFormGroup, UntypedFormBuilder, Validators } from "@angular/forms";
 import { MatTableDataSource } from "@angular/material/table";
 import { ToastrService } from "ngx-toastr";
 import { Observable } from "rxjs";
@@ -20,7 +20,7 @@ export class StockInComponent implements OnInit {
     displayedColumns: string[] = ["barcodeSymbology", "productName", "quantity", "price", "total", "action"];
     products: Product[];
     dataSource = new MatTableDataSource<Product>();
-    orderForm: FormGroup;
+    orderForm: UntypedFormGroup;
     defaultProduct: Product = {
         productId: "",
         quantity: 1,
@@ -51,7 +51,7 @@ export class StockInComponent implements OnInit {
     filteredproducts: any[];
     warehouseLookups: any[];
     constructor(private toastr: ToastrService, 
-        private fb: FormBuilder, 
+        private fb: UntypedFormBuilder, 
         private stockInService: StockInService, 
         private purchaseOrderApi: PurchaseOrderService,
         private warehouseService : WarehouseService) {}
